Доработка моделей. Создание таблиц. Рефакторинг догики админа.

+++ src/main/java/ru/kuleshov/suvinfoservice/menu/handler/MenuHandler.java
@@ -7,6 +7,7 @@ import org.telegram.telegrambots.meta.api.objects.Message;
 import ru.kuleshov.suvinfoservice.bot.TelegramBot;
 import ru.kuleshov.suvinfoservice.service.AbsoluteService;
 import ru.kuleshov.suvinfoservice.state.State;
+import ru.kuleshov.suvinfoservice.utils.Utils;
 
 @Setter
 @Slf4j
@@ -54,27 +55,54 @@ public class MenuHandler {
                 break;
 
             // Обработка меню администратора
-            case WAIT_INPUT_ID_USER_FOR_ADD:
-                absoluteService.getUserService().createUser(Long.parseLong(msg.getText()));
+            case WAIT_INPUT_ID_USER_FOR_ADD: {
+
+                Long userId = Utils.convertsStringToLong(msg.getText());
+                if (userId == null) {
+                    absoluteService.getAction().messageIncorrectInput(msg, bot);
+                    absoluteService.getDaoState().updateState(msg.getChatId(), State.PROCESSING);
+                    break;
+                }
+
+                absoluteService.getUserService().createUser(userId);
 
                 absoluteService.getAction().sendResponse(msg, "Пользователь с telegramId: " +
-                        msg.getText() + " - ДОБАВЛЕН!", bot);
+                        userId + " - ДОБАВЛЕН!", bot);
                 absoluteService.getDaoState().updateState(msg.getChatId(), State.PROCESSING);
                 break;
-            case WAIT_INPUT_ID_USER_FOR_DELETE:
+            }
+            case WAIT_INPUT_ID_USER_FOR_DELETE: {
+
+                Long userId = Utils.convertsStringToLong(msg.getText());
+                if (userId == null) {
+                    absoluteService.getAction().messageIncorrectInput(msg, bot);
+                    absoluteService.getDaoState().updateState(msg.getChatId(), State.PROCESSING);
+                    break;
+                }
+
                 absoluteService.getUserService().deleteUser(Long.parseLong(msg.getText()));
 
                 absoluteService.getAction().sendResponse(msg, "Пользователь с telegramId: " +
                         msg.getText() + " - УДАЛЕН!", bot);
                 absoluteService.getDaoState().updateState(msg.getChatId(), State.PROCESSING);
                 break;
-            case WAIT_INPUT_ID_USER_FOR_DELEGATE_ADMIN_ROOT:
+            }
+            case WAIT_INPUT_ID_USER_FOR_DELEGATE_ADMIN_ROOT: {
+
+                Long userId = Utils.convertsStringToLong(msg.getText());
+                if (userId == null) {
+                    absoluteService.getAction().messageIncorrectInput(msg, bot);
+                    absoluteService.getDaoState().updateState(msg.getChatId(), State.PROCESSING);
+                    break;
+                }
+
                 absoluteService.getUserService().updateUser(Long.parseLong(msg.getText()));
 
                 absoluteService.getAction().sendResponse(msg, "Пользователь с telegramId: " +
                         msg.getText() + " - Наделен правами АДМИНА!", bot);
                 absoluteService.getDaoState().updateState(msg.getChatId(), State.PROCESSING);
                 break;
+            }
         }
     }
 }
